pyproject.toml
[project]
name = "interactive-equity-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
      "agno",
      "openai",
      "yfinance",
      "python-dotenv",
      "sqlalchemy",
      "alembic",
      "pypdf",
      "crawl4ai",
      "pgvector",
      "supabase",
      "psycopg2-binary",
      "ddgs>=9.11.2",
      "psycopg>=3.3.3",
]
